+++ great_expectations/rule_based_profiler/types/data_assistant_result/volume_data_assistant_result.py
@@ -1,4 +1,4 @@
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import Dict, List
 
 import altair as alt
 import pandas as pd
@@ -11,117 +11,10 @@ from great_expectations.rule_based_profiler.types.data_assistant_result import (
 )
 from great_expectations.rule_based_profiler.types.data_assistant_result.plot_result import (
     PlotMode,
-    PlotResult,
 )
 
 
 class VolumeDataAssistantResult(DataAssistantResult):
-    def plot_metrics(
-        self,
-        sequential: bool = True,
-        theme: Optional[Dict[str, Any]] = None,
-        include_column_names: Optional[List[str]] = None,
-        exclude_column_names: Optional[List[str]] = None,
-    ) -> PlotResult:
-        """
-        See parent `DataAssistantResult.plot_metrics()`.
-        """
-        return self._plot(
-            plot_mode=PlotMode.DESCRIPTIVE,
-            sequential=sequential,
-            theme=theme,
-            include_column_names=include_column_names,
-            exclude_column_names=exclude_column_names,
-        )
-
-    def plot_expectations_and_metrics(
-        self,
-        sequential: bool = True,
-        theme: Optional[Dict[str, Any]] = None,
-        include_column_names: Optional[List[str]] = None,
-        exclude_column_names: Optional[List[str]] = None,
-    ) -> PlotResult:
-        """
-        See parent `DataAssistantResult.plot_expectations_metrics()`.
-        """
-        return self._plot(
-            plot_mode=PlotMode.PRESCRIPTIVE,
-            sequential=sequential,
-            theme=theme,
-            include_column_names=include_column_names,
-            exclude_column_names=exclude_column_names,
-        )
-
-    def _plot(
-        self,
-        plot_mode: PlotMode,
-        sequential: bool,
-        theme: Optional[Dict[str, Any]] = None,
-        include_column_names: Optional[List[str]] = None,
-        exclude_column_names: Optional[List[str]] = None,
-    ) -> PlotResult:
-        """
-        VolumeDataAssistant-specific plots are defined with Altair and passed to "display()" for presentation.
-        Display Charts are condensed and interactive while Return Charts are separated into an individual chart for
-        each metric-domain/expectation-domain combination.
-
-        Altair theme configuration reference:
-            https://altair-viz.github.io/user_guide/configuration.html#top-level-chart-configuration
-
-        Args:
-            plot_mode: Type of plot to generate, prescriptive or descriptive
-            sequential: Whether batches are sequential in nature
-            theme: Altair top-level chart configuration dictionary
-            include_column_names: A list of columns to chart
-            exclude_column_names: A list of columns not to chart
-
-        Returns:
-            A PlotResult object consisting of an individual chart for each metric-domain/expectation-domain
-        """
-        if include_column_names is not None and exclude_column_names is not None:
-            raise ValueError(
-                "You may either use `include_column_names` or `exclude_column_names` (but not both)."
-            )
-
-        display_charts: Union[
-            List[alt.Chart], List[alt.LayerChart], List[alt.VConcatChart]
-        ] = []
-        return_charts: Union[List[alt.Chart], List[alt.LayerChart]] = []
-
-        expectation_configurations: List[
-            ExpectationConfiguration
-        ] = self.expectation_configurations
-
-        table_domain_chart: Union[
-            List[alt.Chart], List[alt.LayerChart]
-        ] = self._plot_table_domain_charts(
-            expectation_configurations=expectation_configurations,
-            plot_mode=plot_mode,
-            sequential=sequential,
-        )
-        display_charts.extend(table_domain_chart)
-        return_charts.extend(table_domain_chart)
-
-        column_domain_display_chart: List[alt.VConcatChart]
-        column_domain_return_charts: List[alt.Chart]
-        (
-            column_domain_display_charts,
-            column_domain_return_charts,
-        ) = self._plot_column_domain_charts(
-            expectation_configurations=expectation_configurations,
-            plot_mode=plot_mode,
-            sequential=sequential,
-            include_column_names=include_column_names,
-            exclude_column_names=exclude_column_names,
-        )
-        display_charts.extend(column_domain_display_charts)
-        return_charts.extend(column_domain_return_charts)
-
-        self.display(charts=display_charts, theme=theme)
-
-        return_charts = self.apply_theme(charts=return_charts, theme=theme)
-        return PlotResult(charts=return_charts)
-
     def _create_display_chart_for_column_domain_expectation(
         self,
         expectation_configurations: List[ExpectationConfiguration],
@@ -152,91 +45,3 @@ class VolumeDataAssistantResult(DataAssistantResult):
             plot_mode=plot_mode,
             sequential=sequential,
         )
-
-    def _create_return_chart_for_column_domain_expectation(
-        self,
-        expectation_configuration: ExpectationConfiguration,
-        attributed_metrics: Dict[Domain, Dict[str, ParameterNode]],
-        plot_mode: PlotMode,
-        sequential: bool,
-    ) -> alt.Chart:
-        expectation_metric_map: Dict[str, str] = self.DEFAULT_EXPECTATION_METRIC_MAP
-
-        metric_type: alt.StandardType = AltairDataTypes.QUANTITATIVE.value
-
-        domain: Domain
-        domains_by_column_name: Dict[str, Domain] = {
-            domain.domain_kwargs["column"]: domain
-            for domain in list(attributed_metrics.keys())
-        }
-
-        metric_configuration: dict = expectation_configuration.meta["profiler_details"][
-            "metric_configuration"
-        ]
-        domain_kwargs: dict = metric_configuration["domain_kwargs"]
-
-        domain = domains_by_column_name[domain_kwargs["column"]]
-
-        attributed_values_by_metric_name: Dict[str, ParameterNode] = attributed_metrics[
-            domain
-        ]
-
-        for metric_name in attributed_values_by_metric_name.keys():
-            if (
-                metric_name
-                == expectation_metric_map[expectation_configuration.expectation_type]
-            ):
-                attributed_values: ParameterNode = attributed_values_by_metric_name[
-                    metric_name
-                ]
-
-                df: pd.DataFrame = self._create_df_for_charting(
-                    metric_name=metric_name,
-                    attributed_values=attributed_values,
-                    expectation_configuration=expectation_configuration,
-                    plot_mode=plot_mode,
-                )
-
-                column_name: str = expectation_configuration.kwargs["column"]
-                subtitle = f"Column: {column_name}"
-
-                return self._chart_domain_values(
-                    df=df,
-                    metric_name=metric_name,
-                    metric_type=metric_type,
-                    plot_mode=plot_mode,
-                    sequential=sequential,
-                    subtitle=subtitle,
-                )
-
-    def _chart_column_values(
-        self,
-        column_dfs: List[Tuple[str, pd.DataFrame]],
-        metric_name: str,
-        metric_type: alt.StandardType,
-        plot_mode: PlotMode,
-        sequential: bool,
-    ) -> List[alt.VConcatChart]:
-        plot_impl: Callable[
-            [
-                List[Tuple[str, pd.DataFrame]],
-                str,
-                alt.StandardType,
-            ],
-            alt.VConcatChart,
-        ]
-        if plot_mode is PlotMode.PRESCRIPTIVE:
-            plot_impl = (
-                self.get_interactive_detail_expect_column_values_to_be_between_chart
-            )
-        else:
-            plot_impl = self.get_interactive_detail_multi_chart
-
-        display_chart: alt.VConcatChart = plot_impl(
-            column_dfs=column_dfs,
-            metric_name=metric_name,
-            metric_type=metric_type,
-            sequential=sequential,
-        )
-
-        return [display_chart]
